A tool to generate problem instances for integer bin packing
Problem instances are ordered into different categories and
by the number of qubits in the `data` folder, also added
to gitignore. Note that the tool is probabilistic so there
may be solutions that are not genereated, altough everyone
who runs the unmodified script will generate the same
solutions due to the use of a seed.